backend/requirements.txt
flask
twilio
google-generativeai
python-dotenv
requests
pandas
supabase
openpyxl
